requirements.txt
Flask==2.3.2
flask-socketio==5.3.6
eventlet==0.33.3
pyserial==3.5
